Some more preparation for the upcoming Proxy implementation
git-svn-id: http://reunion.googlecode.com/svn/trunk@288 d5b2e400-7aba-11de-9499-4f8beaa274e3

+++ jreunion/src/com/googlecode/reunion/jreunion/protocol/OtherProtocol.java
@@ -170,7 +170,7 @@ public class OtherProtocol extends Protocol {
 				result = new String(data);
 				break;
 			case 3:
-				short magic6 = (short)(port + 3 * mapId + mapId % 3);
+				short magic6 = (short)((byte)port + 3 * mapId + mapId % 3);
 				for(int i = 0; i < data.length; i++){
 					data[i]=(byte)(((data[i]^magic6)^iterCheck)+4);
 				}
